iptrie: adding RmAll function and making benchmarks more efficient.
To use an iptrie as a cache we need to be able to remove items from time
to time.  Instead of writing a complicated series of function calls to
remove items individually we are providing a way to flush all items.

The benchmarks were taking many minutes to complete, and some were
repetitive, but not as much with this CL.

R=bipasa
CC=dev, gavaletz
https://codereview.appspot.com/14517051